+++ src/entities/signUp/api/postSignup.ts
@@ -1,0 +1,14 @@
+import axios from 'axios';
+
+interface SignUpData {
+  name: string;
+  nickname: string;
+  email: string;
+  password: string;
+  phoneNumber: string;
+}
+
+export const postSignup = async (data: SignUpData) => {
+  const response = await axios.post('/api/auth/signup', data);
+  return response;
+};
